renamed demand_openloop_controller files to have types in the filename
