+++ h2integrate/control/control_strategies/storage/demand_openloop_storage_controller.py
