+++ h2integrate/storage/battery/test/test_battery_cost.py
@@ -4,7 +4,7 @@ import openmdao.api as om
 from pytest import fixture
 
 from h2integrate.storage.battery.atb_battery_cost import ATBBatteryCostModel
-from h2integrate.control.control_strategies.storage.demand_openloop_controller import (
+from h2integrate.control.control_strategies.storage.demand_openloop_storage_controller import (
     DemandOpenLoopStorageController,
 )
